Skip virus tests if clamd installation is new and is in the process of downloading definitions.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@38512 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ virus-base/hier/usr/lib/python2.7/untangle_tests/virus_tests.py
@@ -35,6 +35,7 @@ def nukePassSites():
     node.setPassSites(rules)
 
 class VirusTests(unittest2.TestCase):
+    freshdRunning = False
 
     @staticmethod
     def nodeName():
@@ -69,21 +70,29 @@ class VirusTests(unittest2.TestCase):
 
     # test that client can http download zip
     def test_011_httpNonVirusNotBlocked(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         result = remote_control.runCommand("wget -q -O - http://test.untangle.com/test/test.zip 2>&1 | grep -q text123")
         assert (result == 0)
 
     # test that client can block virus http download zip
     def test_012_httpEicarBlocked(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         result = remote_control.runCommand("wget -q -O - http://test.untangle.com/virus/eicar.zip 2>&1 | grep -q blocked")
         assert (result == 0)
 
     # test that client can block virus http download zip
     def test_013_httpVirusBlocked(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         result = remote_control.runCommand("wget -q -O - http://test.untangle.com/virus/fedexvirus.zip 2>&1 | grep -q blocked")
         assert (result == 0)
 
     # test that client can block virus http download zip
     def test_014_httpEicarPassSite(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         addPassSite("test.untangle.com")
         result = remote_control.runCommand("wget -q -O - http://test.untangle.com/virus/eicar.zip 2>&1 | grep -q blocked")
         nukePassSites()
@@ -91,6 +100,8 @@ class VirusTests(unittest2.TestCase):
 
     # test that client can ftp download zip
     def test_021_ftpNonVirusNotBlocked(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         adResult = subprocess.call(["ping","-c","1","test.untangle.com"],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
         if (adResult != 0):
             raise unittest2.SkipTest("FTP server not available")
@@ -99,6 +110,8 @@ class VirusTests(unittest2.TestCase):
 
     # test that client can block virus ftp download zip
     def test_022_ftpVirusBlocked(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         remote_control.runCommand("rm -f /tmp/temp_022_ftpVirusBlocked_file") 
         result = remote_control.runCommand("wget -q -O /tmp/temp_022_ftpVirusBlocked_file ftp://test.untangle.com/virus/fedexvirus.zip")
         assert (result == 0)
@@ -122,6 +135,8 @@ class VirusTests(unittest2.TestCase):
 
     # test that client can block virus ftp download zip
     def test_023_ftpVirusPassSite(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         ftp_server_IP = socket.gethostbyname("test.untangle.com")
         addPassSite(ftp_server_IP)
         remote_control.runCommand("rm -f /tmp/temp_022_ftpVirusBlocked_file") 
@@ -146,6 +161,8 @@ class VirusTests(unittest2.TestCase):
         nukePassSites()
 
     def test_100_eventlog_httpVirus(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         fname = sys._getframe().f_code.co_name
         result = remote_control.runCommand("wget -q -O - http://test.untangle.com/virus/eicar.zip?arg=%s 2>&1 | grep -q blocked" % fname)
         assert (result == 0)
@@ -163,6 +180,8 @@ class VirusTests(unittest2.TestCase):
         assert( found )
 
     def test_101_eventlog_httpNonVirus(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         fname = sys._getframe().f_code.co_name
         result = remote_control.runCommand("wget -q -O - http://test.untangle.com/test/test.zip?arg=%s 2>&1 | grep -q text123" % fname)
         assert (result == 0)
@@ -180,6 +199,8 @@ class VirusTests(unittest2.TestCase):
         assert( found )
 
     def test_102_eventlog_ftpVirus(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         fname = sys._getframe().f_code.co_name
         result = remote_control.runCommand("wget -q -O /tmp/temp_022_ftpVirusBlocked_file ftp://test.untangle.com/virus/fedexvirus.zip")
         assert (result == 0)
@@ -196,6 +217,8 @@ class VirusTests(unittest2.TestCase):
         assert( found )
 
     def test_103_eventlog_ftpNonVirus(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         fname = sys._getframe().f_code.co_name
         result = remote_control.runCommand("wget -q -O /dev/null ftp://test.untangle.com/test.zip")
         assert (result == 0)
@@ -214,6 +237,8 @@ class VirusTests(unittest2.TestCase):
     port25Test = subprocess.call(["netcat","-z","-w","1","test.untangle.com","25"],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
     @unittest2.skipIf(port25Test != 0,  "Port 25 blocked")
     def test_104_eventlog_smtpVirus(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         startTime = datetime.now()
         fname = sys._getframe().f_code.co_name
         # download the email script
@@ -241,6 +266,8 @@ class VirusTests(unittest2.TestCase):
     port25Test = subprocess.call(["netcat","-z","-w","1","test.untangle.com","25"],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
     @unittest2.skipIf(port25Test != 0,  "Port 25 blocked")
     def test_105_eventlog_smtpNonVirus(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         startTime = datetime.now()
         fname = sys._getframe().f_code.co_name
         result = remote_control.runCommand("echo '%s' > /tmp/attachment-%s" % (fname, fname))
@@ -270,6 +297,8 @@ class VirusTests(unittest2.TestCase):
     port25Test = subprocess.call(["netcat","-z","-w","1","test.untangle.com","25"],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
     @unittest2.skipIf(port25Test != 0,  "Port 25 blocked")
     def test_106_eventlog_smtpVirusPassList(self):
+        if (self.freshdRunning):
+            raise unittest2.SkipTest("clamd is downloading definition files skip all tests")
         ftp_server_IP = socket.gethostbyname("test.untangle.com")
         addPassSite(ftp_server_IP)
         startTime = datetime.now()
